Update patch set 2
Patch Set 2:

What about using the undercloud (or a tester machine).It should be possible to reach the Instances from the undercloud(or a test machine) - the only thing that we need is to write the proper iptables and pass the pybot the right variables. 
What do you think?

Patch-set: 2