Update patch set 2
Patch Set 2: Code-Review-1

> What about using the undercloud (or a tester machine).It should be
 > possible to reach the Instances from the undercloud(or a test
 > machine) - the only thing that we need is to write the proper
 > iptables and pass the pybot the right variables.
 > What do you think?

either is fine for me. putting it on the control node is only because
that's how we do it in upstream deployments. Let's get this to work
(which it's not totally working yet) first and then we can move if
there is a good reason for it.

Itzik, do you know what else I might be missing? in it's current state,
I can't get connectivity from an instance and the 10.10.10.250 address.
seems to be getting dropped on the control node where that ip exists
on br-ex

Patch-set: 2
Label: Code-Review=-1