don't run openstack cleanups before robot by default
Cleanups before running CSIT are only required
when CSIT runs after tests like TEMPEST which
happen to leave plenty of leftover openstack
resources. It's a behaviour in OSP12 while.
In OSP13 we don't have a job that run tempest
and CSIT in one run as well as change the default
to 'not run cleanups' will make running
CSIT quicker 'by default'.

Change-Id: I06eb40d8c0980863ff1861b01738cc466cdf8dd2